requirements.txt
openai==1.3.0
faiss-cpu==1.7.4
langchain==0.0.350
langchain-openai==0.0.2
python-dotenv==1.0.0
streamlit==1.28.0
PyPDF2==3.0.1
python-docx==0.8.11
azure-storage-blob==12.19.0
tiktoken==0.5.1
numpy==1.24.3